requirements.txt
fastapi
pydantic
openai
langgraph
uvicorn
python-dotenv